:recycle: api: DELETE /api/sessions/{id}/certification-candidates/{certificationCandidateId} moved to src
