+++ api/src/certification/session/domain/usecases/delete-unlinked-certification-candidate.js
@@ -1,5 +1,13 @@
-import { CertificationCandidateForbiddenDeletionError } from '../errors.js';
+/**
+ * @typedef {import ('../../../shared/domain/usecases/index.js').dependencies} deps
+ */
 
+import { CertificationCandidateForbiddenDeletionError } from '../../../../../lib/domain/errors.js';
+
+/**
+ * @param {Object} params
+ * @param {deps['certificationCandidateRepository']} params.certificationCandidateRepository
+ */
 const deleteUnlinkedCertificationCandidate = async function ({
   certificationCandidateId,
   certificationCandidateRepository,
